+++ src/transactions/TransactionFrameImpl.cpp
@@ -329,22 +329,31 @@ TransactionFrameImpl::checkValid(Application& app)
         return res;
     }
 
+    bool errorEncountered = false;
+
     AccountRuleVerifierImpl accountRuleVerifier;
     for (auto& op : mOperations)
     {
+        if (errorEncountered)
+        {
+            // we don't what result can return another operation cause changes of previous not applied
+            op->getResult().code(OperationResultCode::opSKIPPED);
+            continue;
+        }
+
         if (!op->checkValid(app, accountRuleVerifier))
         {
-            // it's OK to just fast fail here and not try to call
-            // checkValid on all operations as the resulting object
-            // is only used by applications
-            app.getMetrics()
-                .NewMeter({"transaction", "invalid", "invalid-op"},
-                          "transaction")
-                .Mark();
-            markResultFailed();
-            return false;
+            errorEncountered = true;
+            continue;
         }
     }
+
+    if (errorEncountered)
+    {
+        markResultFailed();
+        return false;
+    }
+
     res = checkAllSignaturesUsed();
     if (!res)
     {